refactor: 입출력 관련 패키지로 이동
