+++ src/test/java/attendance/view/FileReaderTest.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
